Feat : data 객체 생성 & 수정
- RequestUpdatedFixedPlan : membersIds -  Long으로 수정
- ResponseCreateGroup 객체 생성

+++ app/src/main/java/com/kuit/conet/data/dto/request/plan/RequestUpdateFixedPlan.kt
@@ -14,5 +14,5 @@ data class RequestUpdateFixedPlan(
     @SerialName("time")
     val time: String,        // 19:00 형식
     @SerialName("memberIds")
-    val membersIds: List<Int>,
+    val membersIds: List<Long>,
 )
